Bug #2264908 bsp: camera: fix code coverity warning issues
[root cause  ]: some code not comply with specification
[changes     ]:cam_sys
[side effects]:no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums9621
[test case]:camera function
[reviewers]:yangyang.liu

commit_template_version:v1
[change_type ] debug_log
[tag_product ] common

+++ cam_sys/inc/cam_replace_node.h
@@ -54,7 +54,7 @@ enum replace_cmd_type {
 };
 
 struct cam_replace_msg {
-	ssize_t size;
+	uint32_t size;
 	uint32_t align_w;
 	uint32_t align_h;
 	uint32_t offset;
